+++ Plugin/Checkout/Model/ShippingInformationManagementPlugin.php
@@ -10,45 +10,50 @@ declare(strict_types=1);
 
 namespace GLSCroatia\Shipping\Plugin\Checkout\Model;
 
-use GLSCroatia\Shipping\Model\Carrier;
-use Magento\Checkout\Api\Data\ShippingInformationInterface;
-use Magento\Checkout\Api\ShippingInformationManagementInterface;
-
 class ShippingInformationManagementPlugin
 {
+    /**
+     * @var \GLSCroatia\Shipping\Helper\Data
+     */
+    protected \GLSCroatia\Shipping\Helper\Data $dataHelper;
+
     /**
      * @var \Magento\Framework\Serialize\Serializer\Json
      */
     protected \Magento\Framework\Serialize\Serializer\Json $json;
 
     /**
+     * @param \GLSCroatia\Shipping\Helper\Data $dataHelper
      * @param \Magento\Framework\Serialize\Serializer\Json $json
      */
-    public function __construct(\Magento\Framework\Serialize\Serializer\Json $json)
-    {
+    public function __construct(
+        \GLSCroatia\Shipping\Helper\Data $dataHelper,
+        \Magento\Framework\Serialize\Serializer\Json $json
+    ) {
+        $this->dataHelper = $dataHelper;
         $this->json = $json;
     }
 
     /**
      * Save GLS data to the shipping address.
      *
-     * @param ShippingInformationManagementInterface $subject
+     * @param \Magento\Checkout\Api\ShippingInformationManagementInterface $subject
      * @param int $cartId
-     * @param ShippingInformationInterface $addressInformation
+     * @param \Magento\Checkout\Api\Data\ShippingInformationInterface $addressInformation
      * @return void
      * @throws \Magento\Framework\Exception\LocalizedException
      */
     public function beforeSaveAddressInformation(
-        ShippingInformationManagementInterface $subject,
+        \Magento\Checkout\Api\ShippingInformationManagementInterface $subject,
         $cartId,
-        ShippingInformationInterface $addressInformation
+        \Magento\Checkout\Api\Data\ShippingInformationInterface $addressInformation
     ): void {
         $shippingAddress = $addressInformation->getShippingAddress();
 
         $glsData = $this->jsonDecode($shippingAddress->getData('gls_data') ?: '');
 
-        $isParcelShopDelivery = $addressInformation->getShippingCarrierCode() === Carrier::CODE
-            && $addressInformation->getShippingMethodCode() === Carrier::PARCEL_SHOP_DELIVERY_METHOD;
+        $isParcelShopDelivery = $addressInformation->getShippingCarrierCode() === \GLSCroatia\Shipping\Model\Carrier::CODE // phpcs:ignore
+            && $this->dataHelper->isLockerShopDeliveryMethod($addressInformation->getShippingMethodCode());
 
         $deliveryPoint = $this->jsonDecode(
             $addressInformation->getExtensionAttributes()->getGlsParcelShopDeliveryPoint() ?: ''